+++ app/Http/Controllers/Settings/SubscriptionsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Settings;
 
 use Illuminate\View\View;
+use App\Traits\StripeCall;
 use App\Helpers\DateHelper;
 use Illuminate\Http\Request;
 use Laravel\Cashier\Cashier;
@@ -21,6 +22,8 @@ use App\Services\Account\Settings\ArchiveAllContacts;
 
 class SubscriptionsController extends Controller
 {
+    use StripeCall;
+
     /**
      * Display a listing of the resource.
      *
@@ -96,10 +99,16 @@ class SubscriptionsController extends Controller
      */
     public function confirmPayment($id)
     {
+        try {
+            $payment = $this->stripeCall(function () use ($id) {
+                return StripePaymentIntent::retrieve($id, Cashier::stripeOptions());
+            });
+        } catch (StripeException $e) {
+            return back()->withErrors($e->getMessage());
+        }
+
         return view('settings.subscriptions.confirm', [
-            'payment' => new Payment(
-                StripePaymentIntent::retrieve($id, Cashier::stripeOptions())
-            ),
+            'payment' => new Payment($payment),
             'redirect' => request('redirect'),
         ]);
     }
